refactor: #101 dateToYmdSlash util 함수로 변경
