+++ utils/dateToYmdSlash.ts
@@ -1,0 +1,7 @@
+export const dateToYmdSlash = (date: string): string => {
+  const parsedDate = new Date(date)
+  const year = parsedDate.getFullYear()
+  const month = String(parsedDate.getMonth() + 1).padStart(2, "0")
+  const day = String(parsedDate.getDate()).padStart(2, "0")
+  return `${year}/${month}/${day}`
+}
